refactor: 위치 기반 관광 정보 조회 서비스 인터페이스 분리
- 서비스 구현체로부터 인터페이스를 분리
